% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_model_output.R
\name{aggregate_model_output}
\alias{aggregate_model_output}
\title{create some aggregates of various columns in the output dataframe from the model}
\usage{
aggregate_model_output(model, output)
}
\arguments{
\item{model}{model object}

\item{output}{model output}
}
\value{
aggregated data
}
\description{
create some aggregates of various columns in the output dataframe from the model
}
