% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annealing_fit_ecology_parameters.R
\name{annealing_fit_ecology_parameters}
\alias{annealing_fit_ecology_parameters}
\title{Finds the best-fit values of the ecology model parameters using an annealing process}
\usage{
annealing_fit_ecology_parameters(model)
}
\arguments{
\item{model}{current model configuration}
}
\value{
fitted parameter set
}
\description{
Finds the best-fit values of the ecology model parameters using an annealing process
}
