% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_flow_matrix_from_model_annual_output.R
\name{assemble_flow_matrix_from_model_annual_output}
\alias{assemble_flow_matrix_from_model_annual_output}
\title{process the whole domain annual flux results held in the dataframe annual_flux_results
to produce a flow-matrix defining all of the annual integrated mass fluxes been every pair of
prey and predator guilds in the food web, alf the geochemical flows, and all of the boundary flows}
\usage{
assemble_flow_matrix_from_model_annual_output(model, build, output,
  aggregates)
}
\arguments{
\item{model}{model object}

\item{build}{model build object}

\item{output}{model output}

\item{aggregates}{aggregated model output}
}
\description{
The code then uses the NetIndices R package which relies on the flow-matrix data as input
to derive a suite of network indices as outputs These are saved as a standard csv file
}
