% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_and_whisker_annual_plots.R
\name{box_and_whisker_annual_plots}
\alias{box_and_whisker_annual_plots}
\title{plot observed and modelled distributions of the annual ecosystem flux data}
\usage{
box_and_whisker_annual_plots(model, annual.target.data, opt_results)
}
\arguments{
\item{model}{model object}

\item{annual.target.data}{annual target data}

\item{opt_results}{model target results}
}
\description{
plot observed and modelled distributions of the annual ecosystem flux data
}
