% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_and_whisker_monthly_plots.R
\name{box_and_whisker_monthly_plots}
\alias{box_and_whisker_monthly_plots}
\title{plot some monthly results}
\usage{
box_and_whisker_monthly_plots(model)
}
\arguments{
\item{model}{model object}
}
\description{
plot some monthly results
}
