% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_annual_drivers.R
\name{build_annual_drivers}
\alias{build_annual_drivers}
\title{uses monthly values of time-series drivers over a climatological annual cycle
and makes them into functions for #interpolating values at any given time.}
\usage{
build_annual_drivers(run, fixed.parms, physical.parms, physics.drivers,
  chemistry.drivers, biological.events)
}
\arguments{
\item{run}{model run object}

\item{fixed.parms}{fixed parameters}

\item{physical.parms}{physical parameters}

\item{physics.drivers}{physics drivers}

\item{chemistry.drivers}{chemistry drivers}

\item{biological.events}{biological timing events}
}
\value{
driver functions
}
\description{
returns set of interpolating functions
}
