% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fishing_fleet_model.R
\name{build_fishing_fleet_model}
\alias{build_fishing_fleet_model}
\title{build fishing fleet model taking notice of any changes to physical depth parameters}
\usage{
build_fishing_fleet_model(fleet.model, physical.parms)
}
\arguments{
\item{model.path}{path to users model folder, otherwise read package model}
}
\value{
model object
}
\description{
build fishing fleet model taking notice of any changes to physical depth parameters
}
