% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_initial_state.R
\name{build_initial_state}
\alias{build_initial_state}
\title{adjust initial state taking notice of physical parameters}
\usage{
build_initial_state(initial.state, physical.parms)
}
\arguments{
\item{initial.state}{initial state variables}

\item{physical.parms}{model physical parameters}
}
\value{
adjusted initial state variables
}
\description{
adjust initial state taking notice of physical parameters
}
