% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{build model for invocation of StrathE2E}
\usage{
build_model(model, nyears)
}
\arguments{
\item{model}{model object}
}
\value{
build object
}
\description{
calculates derived variables such as annual drivers, forcings, fleet model, uptakes, etc.
}
