% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_parameters.R
\name{build_model_parameters}
\alias{build_model_parameters}
\title{Build entire set of model parameters combining fixed, physical, fleet model output, plus some biomass thresholds}
\usage{
build_model_parameters(model, fleet.model.output)
}
\arguments{
\item{model}{model object}

\item{fleet.model.output}{fleet model output}
}
\value{
model parameter vector
}
\description{
Build entire set of model parameters combining fixed, physical, fleet model output, plus some biomass thresholds
}
