% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_run.R
\name{build_model_run}
\alias{build_model_run}
\title{build model run}
\usage{
build_model_run(nyears)
}
\arguments{
\item{nyears}{number of years to run the model}
}
\value{
model run object
}
\description{
build run time information from number of years
}
