% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_error_function.R
\name{calculate_error_function}
\alias{calculate_error_function}
\title{measure current model fit}
\usage{
calculate_error_function(model, opt_results)
}
\arguments{
\item{model}{current model configuration}

\item{opt_results}{model target data}
}
\value{
mean squared errors
}
\description{
error function for fitting process
}
