% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uptakes.R
\name{calculate_uptakes}
\alias{calculate_uptakes}
\title{Apply perturbed preference values to fitted parameters to calculate uptakes}
\usage{
calculate_uptakes(fitted.parms)
}
\arguments{
\item{fitted.parms}{fitted model parameters}
}
\value{
uptake parameters
}
\description{
Apply perturbed preference values to fitted parameters to calculate uptakes
}
