% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_fishing_fleet_model.R
\name{configure_fishing_fleet_model}
\alias{configure_fishing_fleet_model}
\title{configure fishing fleet model
reads model setup from various configuration files and configures the model}
\usage{
configure_fishing_fleet_model(model.path, physical.parms)
}
\arguments{
\item{model.path}{path to users model folder, otherwise read package model}

\item{physical.parms}{physical model parameters}
}
\value{
model object
}
\description{
configure fishing fleet model
reads model setup from various configuration files and configures the model
}
