% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_model.R
\name{copy_model}
\alias{copy_model}
\title{make a copy of the named model/variant}
\usage{
copy_model(model.name, model.variant, dest.path = "Models",
  user.path = "")
}
\arguments{
\item{model.name}{name of model to copy}

\item{model.variant}{name of model variant to copy}

\item{dest.path}{destination folder to write copy of model, will be created if necessary, default is "Models" in current folder}

\item{user.path}{path to users top level model folder if copying a user model}
}
\description{
make a copy of the named model/variant
}
