% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_annual_results_inshore.R
\name{derive_annual_results_inshore}
\alias{derive_annual_results_inshore}
\title{derive set of results for inshore only and write to files}
\usage{
derive_annual_results_inshore(model, results)
}
\arguments{
\item{model}{model object}

\item{results}{model results}
}
\description{
derive set of results for inshore only and write to files
}
