% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_annual_results_offshore.R
\name{derive_annual_results_offshore}
\alias{derive_annual_results_offshore}
\title{derive set of results for offshore only and write to files}
\usage{
derive_annual_results_offshore(model, build, output, aggregates)
}
\arguments{
\item{model}{model object}

\item{build}{model build object}

\item{output}{model output}

\item{aggregates}{aggregated model output}
}
\description{
derive set of results for offshore only and write to files
}
