% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_annual_results_wholedomain.R
\name{derive_annual_results_wholedomain}
\alias{derive_annual_results_wholedomain}
\title{extract annual average biomass in the final year for the whole model domain}
\usage{
derive_annual_results_wholedomain(model, results)
}
\arguments{
\item{model}{model object}

\item{results}{model results}
}
\description{
extract annual average biomass in the final year for the whole model domain
}
