% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_model_target_results.R
\name{derive_model_target_results}
\alias{derive_model_target_results}
\title{derive annual target results and write to file}
\usage{
derive_model_target_results(model, output, aggregates, annualtargetdata)
}
\arguments{
\item{model}{model object}

\item{output}{model output}

\item{aggregates}{aggregated model output}

\item{annualtargetdata}{annual target data}
}
\value{
target results
}
\description{
returns annual target results
}
