% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregate_landings_discards_by_gear.R
\name{disaggregate_landings_discards_by_gear}
\alias{disaggregate_landings_discards_by_gear}
\title{disaggregate group annual landings and discards by gear}
\usage{
disaggregate_landings_discards_by_gear(fleet.model.output, annual.landings)
}
\arguments{
\item{fleet.model.output}{fleet model output}

\item{annual.landings}{annual landings}
}
\value{
model output with gear disaggregation
}
\description{
returns a model object with run and data slots
}
