% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e2e_info_pk.R
\name{e2e_info_pk}
\alias{e2e_info_pk}
\title{List the version mumber and details of the installed StrathE2E2 package.}
\usage{
e2e_info_pk(pk = "StrathE2E2")
}
\arguments{
\item{pk}{Package name. Default="StrathE2E2"}
}
\value{
An on-screen list of the installed package details.
}
\description{
List the version mumber and details of the installed StrathE2E2 package.
}
\examples{
# List the version and details of the package:
#    e2e_info_pk()

}
\seealso{
\code{\link{e2e_copy}}, \code{\link{e2e_ls}} , \code{\link{e2e_info_md}}
}
