% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e2e_ls.R
\name{e2e_ls}
\alias{e2e_ls}
\title{List the available models in a designated workspace.}
\usage{
e2e_ls(models.path = NULL)
}
\arguments{
\item{models.path}{Relative path from the current working directory to a folder containing a library of model configurations (typically "Folder/Models"). Setting models.path="" is valid. Default models.path=NULL, meaning read a North Sea model setup from the package folder extdata/Models.}
}
\value{
An on-screen list of available models in the designated folder.
}
\description{
List the available models in a designated workspace.
}
\examples{
# List the models/variants supplied with the package:
    e2e_ls()

# Dummy example to illustrate listing the models/variants in a user defined
# workspace. REPLACE "Folder1/Models" with your own models.path 
# remembering that this are relative to the current working directory.
#    e2e_ls("Folder/Models")

}
\seealso{
\code{\link{e2e_copy}}, \code{\link{e2e_read}}
}
