% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/extract_simulated_catch_land_disc_by_gear_for_given_year.R
\name{extract_simulated_catch_land_disc_by_gear_for_given_year}
\alias{extract_simulated_catch_land_disc_by_gear_for_given_year}
\title{extract simulated catch}
\usage{
extract_simulated_catch_land_disc_by_gear_for_given_year(model, results)
}
\arguments{
\item{model}{model object}

\item{results}{model results}
}
\description{
saves to various files
}
