% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_fleet_model.R
\name{fishing_fleet_model}
\alias{fishing_fleet_model}
\title{defines the fishing fleet model}
\usage{
fishing_fleet_model(model, build)
}
\arguments{
\item{model}{model object}

\item{build}{model build object}
}
\value{
fishing fleet model
}
\description{
returns fishing fleet model
}
