% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_drivers.R
\name{interpolate_drivers}
\alias{interpolate_drivers}
\title{Interpolate drivers over time that model is to be run}
\usage{
interpolate_drivers(run, drivers)
}
\arguments{
\item{run}{model run settings}

\item{drivers}{model drivers}
}
\value{
model output
}
\description{
Interpolate drivers over time that model is to be run
}
