% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{list available models}
\usage{
list_models(model.path = "")
}
\arguments{
\item{model.path}{path to users model folder, otherwise read package model folder}
}
\description{
list available models
}
