% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_averages_of_final_year.R
\name{monthly_averages_of_final_year}
\alias{monthly_averages_of_final_year}
\title{Summarise the final year for nitrate ammonia and chl}
\usage{
monthly_averages_of_final_year(model, build, output, aggregates)
}
\arguments{
\item{model}{model object}

\item{build}{model build object}

\item{output}{model output}

\item{aggregates}{aggregated model output}
}
\value{
monthly averages
}
\description{
Summarise the final year for nitrate ammonia and chl
}
