% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb_parameters.R
\name{perturb_parameters}
\alias{perturb_parameters}
\title{perturb parameters a bit}
\usage{
perturb_parameters(datastore, annealing.parms = list())
}
\arguments{
\item{datastore}{parameters to be perturbed}

\item{annealing.parms}{annealing paramters}
}
\value{
perturbed parameter vector
}
\description{
perturb parameters a bit
}
