% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_final_year.R
\name{plot_final_year}
\alias{plot_final_year}
\title{Plot final year time series}
\usage{
plot_final_year(model, results)
}
\arguments{
\item{model}{model object}

\item{results}{full model results}
}
\description{
Plot final year time series
}
