% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_final_year_sediment_habitats.R
\name{plot_final_year_sediment_habitats}
\alias{plot_final_year_sediment_habitats}
\title{Plot full time series of out}
\usage{
plot_final_year_sediment_habitats(model, results)
}
\arguments{
\item{model}{model object}

\item{results}{full model results}
}
\description{
Plot set of time series from model out
}
