% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_for_each_gear_catch_by_group.R
\name{plot_for_each_gear_catch_by_group}
\alias{plot_for_each_gear_catch_by_group}
\title{plot results as lineplots by group}
\usage{
plot_for_each_gear_catch_by_group(model, results)
}
\arguments{
\item{model}{model object}

\item{results}{model results object}
}
\description{
plot results as lineplots by group
}
