% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sediment_habitats_final_year.R
\name{plot_sediment_habitats_final_year}
\alias{plot_sediment_habitats_final_year}
\title{Plot full time series of out}
\usage{
plot_sediment_habitats_final_year(model, results)
}
\arguments{
\item{model}{model object}

\item{results}{full model results}
}
\description{
Plot set of time series from model out
}
