% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series_annual_land_disc.R
\name{plot_time_series_annual_land_disc}
\alias{plot_time_series_annual_land_disc}
\title{plot results as lineplots by group}
\usage{
plot_time_series_annual_land_disc(model, results)
}
\arguments{
\item{model}{model object}

\item{results}{model results object}
}
\description{
plot results as lineplots by group
}
