% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annealing_parameters.R
\name{read_annealing_parameters}
\alias{read_annealing_parameters}
\title{return list of annealing paramters}
\usage{
read_annealing_parameters(model.path)
}
\arguments{
\item{model.path}{model path}
}
\value{
list of annealing parameters
}
\description{
return list of annealing paramters
}
