% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annual_target_data.R
\name{read_annual_target_data}
\alias{read_annual_target_data}
\title{read annual target data}
\usage{
read_annual_target_data(model.path)
}
\arguments{
\item{model.path}{path to current model/version folder}
}
\value{
target data
}
\description{
returns target data
}
