% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biological_event_timings.R
\name{read_biological_event_timings}
\alias{read_biological_event_timings}
\title{Read the biological event timing parameters}
\usage{
read_biological_event_timings(model.path)
}
\arguments{
\item{model.path}{path to model}
}
\value{
events
}
\description{
returns event timings
}
