% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemistry_drivers.R
\name{read_chemistry_drivers}
\alias{read_chemistry_drivers}
\title{Read the default set of chemistry drivers}
\usage{
read_chemistry_drivers(model.path)
}
\arguments{
\item{model.path}{path to model}
}
\value{
chemistry driver set
}
\description{
returns drivers in a named list
}
