% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fitted_parameters.R
\name{read_fitted_parameters}
\alias{read_fitted_parameters}
\title{Load fitted parameter values for model}
\usage{
read_fitted_parameters(model.path)
}
\arguments{
\item{model.path}{path to model}
}
\value{
list of fitted model parameters
}
\description{
Parameters which have been established by the simulated annealing scheme are
loaded from a csv file which is made from the last line of the 'accepted parameters'
file produced by the annealing programme.
A variety of other parameters which are not optimised by simulated annealing are
hard-wired in this subroutine
}
