% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fixed_parameters.R
\name{read_fixed_parameters}
\alias{read_fixed_parameters}
\title{load the fixed parameters used in the model}
\usage{
read_fixed_parameters(model.path)
}
\arguments{
\item{model.path}{path to model}
}
\value{
fixed model parameters
}
\description{
load the fixed parameters used in the model
}
