% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_initial_state.R
\name{read_initial_state}
\alias{read_initial_state}
\title{read and set the initial values of the state variables used in the model}
\usage{
read_initial_state(model.path, physical.parms)
}
\arguments{
\item{model.path}{path to model files}

\item{physical.parms}{model physical parameters}
}
\value{
initial state variables
}
\description{
Initial values are set by specifying the initial surface water nitrate concentration, and
then the ratios of all the other state variables to this value.
}
