% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_physical_parameters.R
\name{read_physical_parameters}
\alias{read_physical_parameters}
\title{set up the physical configuration of the model}
\usage{
read_physical_parameters(model.path)
}
\arguments{
\item{model.path}{path to model}
}
\value{
default physical parameters
}
\description{
Data required are the vertical thicknesses of the surface, deep and sediment layer,
sediment porosity, and the diffusion coefficients for exchange between the sediment and water
The parameters are read from the "physical_parameters.csv" file.
}
