% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_physics_drivers.R
\name{read_physics_drivers}
\alias{read_physics_drivers}
\title{Read the default set of physics drivers}
\usage{
read_physics_drivers(model.path)
}
\arguments{
\item{model.path}{path to model}
}
\value{
physics driver set
}
\description{
returns drivers in a named list
}
