% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SD_control_ecology.R
\name{read_SD_control_ecology}
\alias{read_SD_control_ecology}
\title{return list of annealing paramters}
\usage{
read_SD_control_ecology(model.path)
}
\arguments{
\item{model.path}{model path}
}
\value{
list of annealing parameters for fitting ecology model
}
\description{
return list of annealing paramters
}
