% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_model.R
\name{rebuild_model}
\alias{rebuild_model}
\title{rebuild model object}
\usage{
rebuild_model(model)
}
\arguments{
\item{model}{existing model object}
}
\value{
rebuilt model object
}
\description{
restore consistency to model data after modifying base parameter set
}
