% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_model_data.R
\name{rebuild_model_data}
\alias{rebuild_model_data}
\title{rebuild model data object}
\usage{
rebuild_model_data(model)
}
\arguments{
\item{model}{existing model object}
}
\value{
model with rebuilt data object
}
\description{
restore consistency to model data after modifying base parameter set
}
