% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_model_run.R
\name{rebuild_model_run}
\alias{rebuild_model_run}
\title{rebuild model run}
\usage{
rebuild_model_run(model)
}
\arguments{
\item{model}{existing model}
}
\value{
model with rebuilt run
}
\description{
restore consistency to model run if nyears has been modified
}
