% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{performs a one-at-a-time sensitivity analysis using
the Morris Method for factorial sampling of the ecology model
parameters, the fishing fleet model parameters, and the
environmental forcings}
\usage{
sensitivity_analysis(model)
}
\arguments{
\item{model}{full model object}
}
\value{
list of default model run settings
}
\description{
performs a one-at-a-time sensitivity analysis using
the Morris Method for factorial sampling of the ecology model
parameters, the fishing fleet model parameters, and the
environmental forcings
}
