% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default_run.R
\name{set_default_run}
\alias{set_default_run}
\title{return settings for a default run}
\usage{
set_default_run(model.name, model.variant, model.subdir = "",
  model.ident = "base", nyears = 20)
}
\arguments{
\item{model.name}{name of model being loaded}

\item{model.variant}{which model variant being loaded}

\item{model.subdir}{results written to this sub-directory}

\item{model.ident}{appended to output filenames, default "base"}

\item{nyears}{length of model run in years}
}
\value{
list of default model run settings
}
\description{
returns named list of time settings, output directories
}
