% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_fleet_model_output.R
\name{store_fleet_model_output}
\alias{store_fleet_model_output}
\title{grab the 'gear_group_prop_effort' and 'gear_group_prob_discard' matrices, reorganise and store them
for use after the main model run to derive the landings and discards from each gear, from inshore and offshore zones}
\usage{
store_fleet_model_output(fleet_model_output, years_to_fill)
}
\arguments{
\item{fleet_model_output}{output from fleet model}

\item{years_to_fill}{makes data fill the run length}
}
\value{
reorganised fleet model output
}
\description{
returns fishing fleet model output suitably reorganised
}
