% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrathE2E_anneal.R
\name{StrathE2E_anneal}
\alias{StrathE2E_anneal}
\title{Attempt to fit the model to the given data}
\usage{
StrathE2E_anneal(model)
}
\arguments{
\item{model}{current model configuration}
}
\value{
fitted parameter set
}
\description{
Uses a simulated annealing process to fit the model data
}
\seealso{
\code{\link{list_models}}, \code{\link{read_model}}, \code{\link{copy_model}}, \code{\link{rebuild_model}}
}
