% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fitted_parameters.R
\name{write_fitted_parameters}
\alias{write_fitted_parameters}
\title{Save current set of fitted parameters to file}
\usage{
write_fitted_parameters(model, parhistory)
}
\arguments{
\item{parhistory}{parameter set history}

\item{model.path}{path to model}
}
\value{
list of fitted model parameters
}
\description{
Save current set of fitted parameters to file
}
