% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_and_whisker_monthly_plots.R
\name{box_and_whisker_monthly_plots}
\alias{box_and_whisker_monthly_plots}
\title{plot some monthly result sumamries}
\usage{
box_and_whisker_monthly_plots(results)
}
\arguments{
\item{results}{model results object}
}
\description{
plot some monthly result sumamries
}
