% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_model.R
\name{copy_model}
\alias{copy_model}
\title{make a copy of the named model (and all it's variants)}
\usage{
copy_model(model.name, dest.path = "Models", user.path = "",
  overwrite = FALSE)
}
\arguments{
\item{model.name}{name of model to copy}

\item{dest.path}{destination folder to write copy of model, will be created if necessary, default is "Models" in current folder}

\item{user.path}{path to users top level model folder if copying a user model}

\item{overwrite}{set to TRUE to enable overwrite of existing model folder}
}
\description{
make a copy of the named model (and all it's variants)
}
