% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_model_target_results.R
\name{derive_model_target_results}
\alias{derive_model_target_results}
\title{derive annual target results and write to file (ZZ why?)}
\usage{
derive_model_target_results(model, results, annualtargetdata)
}
\arguments{
\item{model}{model object}

\item{results}{model results}

\item{annualtargetdata}{annual target data}
}
\value{
target results
}
\description{
returns annual target results
}
