% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_timeseries_annual_landings.R
\name{extract_timeseries_annual_landings}
\alias{extract_timeseries_annual_landings}
\title{read designated model}
\usage{
extract_timeseries_annual_landings(model, out)
}
\arguments{
\item{model}{current model}

\item{out}{model output}
}
\value{
inshore/offshore annual landings
}
\description{
returns a model object with run and data slots
}
