% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_averages_of_final_year.R
\name{monthly_averages_of_final_year}
\alias{monthly_averages_of_final_year}
\title{Summarise the final year for nitrate ammonia and chl and write to a file}
\usage{
monthly_averages_of_final_year(model, results)
}
\arguments{
\item{model}{model object}

\item{results}{model results}
}
\description{
Summarise the final year for nitrate ammonia and chl and write to a file
}
