% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_for_each_group_catch_by_gear.R
\name{plot_for_each_group_catch_by_gear}
\alias{plot_for_each_group_catch_by_gear}
\title{plot results as lineplots by group}
\usage{
plot_for_each_group_catch_by_gear(model, catch.land.disc)
}
\arguments{
\item{model}{model object}

\item{catch.land.disc}{catch, landings and discards}
}
\description{
plot results as lineplots by group
}
