% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fishing_fleet_model.R
\name{read_fishing_fleet_model}
\alias{read_fishing_fleet_model}
\title{read fishing fleet model configuration}
\usage{
read_fishing_fleet_model(model.path, physical.parameters)
}
\arguments{
\item{model.path}{path to users model folder, otherwise read package model}

\item{physical.parameters}{physical model parameters}
}
\value{
model object
}
\description{
read fishing fleet model configuration
}
