% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrathE2E.R
\name{StrathE2E}
\alias{StrathE2E}
\title{run the StrathE2E model}
\usage{
StrathE2E(model)
}
\arguments{
\item{model}{current model configuration}
}
\value{
model output
}
\description{
perform a single run of the StrathE2E model
}
\examples{
# after running the example below, look at the various objects to see how you can modify them:
#
#	str(NorthSea)
#
NorthSea <- read_model("NorthSea", "version1")

results <- StrathE2E(NorthSea)

plot_full_length_timeseries(results)
}
\seealso{
\code{\link{list_models}}, \code{\link{read_model}}, \code{\link{copy_model}}, \code{\link{rebuild_model}}
}
